Prevent destination folder from being deleted
There is a major issue when delete an entry that contains an empty zip upload
field: The destination folder would get deleted.

This commits prevents that, by checking that we have a valide
destination and a valid file path before returning the file location
array.

Each call site much check for null returns.